Move download feature to a separate component
