+++ components/index.js
@@ -3,3 +3,4 @@ export * from './options-box/options-box'
 export * from './result-table/result-table'
 export * from './footer/footer'
 export * from './top-bar/top-bar'
+export * from './download-results/download-results'
